Refactor base URL utility: use getBaseUrl.ts everywhere, remove baseUrl.ts, update tests and imports.
