+++ src/components/AppHeader.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import AppHeader from './AppHeader';
-jest.mock('../utils/baseUrl', () => ({ BASE_URL: '/' }));
+jest.mock('../utils/getBaseUrl', () => ({ getBaseUrl: () => '/' }));
 
 describe('AppHeader', () => {
   it('renders the app icon with correct alt text', () => {
