+++ src/pages/Achievements.test.tsx
@@ -23,7 +23,7 @@ jest.mock('../badges', () => ({
 }));
 
 // Mock BASE_URL
-jest.mock('../utils/baseUrl', () => ({ BASE_URL: '/' }));
+jest.mock('../utils/getBaseUrl', () => ({ getBaseUrl: () => '/' }));
 
 // Mock Modal to just render children for simplicity
 jest.mock('../components/Modal', () => ({
